rename jsondatamodel.ts -> jsonmodel.ts
