+++ packages/datagrid/src/index.ts
@@ -9,5 +9,5 @@ export * from './cellrenderer';
 export * from './datagrid';
 export * from './datamodel';
 export * from './graphicscontext';
-export * from './jsondatamodel';
+export * from './jsonmodel';
 export * from './textrenderer';
